pyproject.toml
[project]
name = "bustle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-phonenumber-field>=8.4.0",
    "dotenv>=0.9.9",
    "phonenumbers>=9.0.23",
    "pillow>=12.1.0",
]
